add autonat to behaviour
