+++ Cargo.toml
@@ -26,7 +26,7 @@ hyper = "0.14" # Prometheus server
 iri-string = "0.5"
 lazy_static = "1.4.0"
 libipld = "0.14"
-libp2p = { default-features = false, features = ["gossipsub", "identify", "kad", "tcp-tokio", "mplex", "noise", "ping", "yamux", "dns-tokio", "relay", "dcutr"], version = "0.49.0" }
+libp2p = { default-features = false, features = ["autonat", "gossipsub", "identify", "kad", "tcp-tokio", "mplex", "noise", "ping", "yamux", "dns-tokio", "relay", "dcutr"], version = "0.49.0" }
 opentelemetry = { version = "0.17.0", features = ["rt-tokio"] }
 opentelemetry-jaeger = { version = "0.16.0", features = ["rt-tokio", "reqwest_collector_client"] }
 pin-project = "1"
